requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
python-docx>=1.1.0
pypdf>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
validators>=0.22.0
openai>=1.10.0
Pillow>=10.2.0